+++ src/simpy_demo/expressions/__init__.py
@@ -1,5 +1,0 @@
-"""Expression engine for evaluating YAML config expressions."""
-
-from simpy_demo.expressions.engine import ExpressionEngine
-
-__all__ = ["ExpressionEngine"]