+++ src/simpy_demo/factories/__init__.py
@@ -1,5 +1,0 @@
-"""Factory classes for generating simulation entities."""
-
-from simpy_demo.factories.telemetry import TelemetryGenerator
-
-__all__ = ["TelemetryGenerator"]